+++ .coveragerc
@@ -317,6 +317,8 @@ omit =
     homeassistant/components/knx/*
     homeassistant/components/knx/climate.py
     homeassistant/components/knx/cover.py
+    homeassistant/components/kodi/__init__.py
+    homeassistant/components/kodi/const.py
     homeassistant/components/kodi/media_player.py
     homeassistant/components/kodi/notify.py
     homeassistant/components/konnected/*
