+++ homeassistant/components/kodi/const.py
@@ -1,0 +1,2 @@
+"""Constants for the Kodi platform."""
+DOMAIN = "kodi"
