+++ homeassistant/components/media_player/services.yaml
@@ -232,35 +232,6 @@ soundtouch_remove_zone_slave:
       description: Name of slaves entities to remove from the existing zone.
       example: 'media_player.soundtouch_bedroom'
 
-kodi_add_to_playlist:
-  description: Add music to the default playlist (i.e. playlistid=0).
-  fields:
-    entity_id:
-      description: Name(s) of the Kodi entities where to add the media.
-      example: 'media_player.living_room_kodi'
-    media_type:
-      description: Media type identifier. It must be one of SONG or ALBUM.
-      example: ALBUM
-    media_id:
-      description: Unique Id of the media entry to add (`songid` or albumid`). If not defined, `media_name` and `artist_name` are needed to search the Kodi music library.
-      example: 123456
-    media_name:
-      description: Optional media name for filtering media. Can be 'ALL' when `media_type` is 'ALBUM' and `artist_name` is specified, to add all songs from one artist.
-      example: 'Highway to Hell'
-    artist_name:
-      description: Optional artist name for filtering media.
-      example: 'AC/DC'
-
-kodi_call_method:
-  description: 'Call a Kodi JSONRPC API method with optional parameters. Results of the Kodi API call will be redirected in a Home Assistant event: `kodi_call_method_result`.'
-  fields:
-    entity_id:
-      description: Name(s) of the Kodi entities where to run the API method.
-      example: 'media_player.living_room_kodi'
-    method:
-      description: Name of the Kodi JSONRPC API method to be called.
-      example: 'VideoLibrary.GetRecentlyAddedEpisodes'
-
 squeezebox_call_method:
   description: 'Call a Squeezebox JSON/RPC API method.'
   fields:
